+++ lib/pub_sub_model_sync.rb
@@ -5,6 +5,7 @@ require 'active_support'
 
 require 'pub_sub_model_sync/railtie'
 require 'pub_sub_model_sync/config'
+require 'pub_sub_model_sync/base'
 require 'pub_sub_model_sync/subscriber_concern'
 require 'pub_sub_model_sync/message_publisher'
 require 'pub_sub_model_sync/publisher_concern'