+++ lib/pub_sub_model_sync/config.rb
@@ -5,16 +5,27 @@ module PubSubModelSync
     cattr_accessor(:subscribers) { [] }
     cattr_accessor(:publishers) { [] }
     cattr_accessor(:service_name) { :google }
-    cattr_accessor :logger
+
+    # customizable callbacks
+    cattr_accessor(:debug) { false }
+    cattr_accessor :logger # LoggerInst
+
+    cattr_accessor(:on_before_processing) { ->(_payload, _subscriber) {} } # return :cancel to skip
+    cattr_accessor(:on_success_processing) { ->(_payload, _subscriber) {} }
+    cattr_accessor(:on_error_processing) { ->(_exception, _payload) {} }
+    cattr_accessor(:on_before_publish) { ->(_payload) {} } # return :cancel to skip
+    cattr_accessor(:on_after_publish) { ->(_payload) {} }
+    cattr_accessor(:on_error_publish) { ->(_exception, _payload) {} }
+    cattr_accessor(:disabled_callback_publisher) { ->(_model, _action) { false } }
 
     # google service
     cattr_accessor :project, :credentials, :topic_name, :subscription_name
 
     # rabbitmq service
-    cattr_accessor :bunny_connection, :queue_name, :topic_name
+    cattr_accessor :bunny_connection, :queue_name, :topic_name, :subscription_name
 
     # kafka service
-    cattr_accessor :kafka_connection, :topic_name
+    cattr_accessor :kafka_connection, :topic_name, :subscription_name
 
     def self.log(msg, kind = :info)
       msg = "PS_MSYNC ==> #{msg}"
@@ -23,6 +34,11 @@ module PubSubModelSync
       else
         logger ? logger.send(kind, msg) : puts(msg)
       end
+    end
+
+    def self.subscription_key
+      subscription_name ||
+        (Rails.application.class.parent_name rescue '') # rubocop:disable Style/RescueModifier
     end
   end
 end