+++ lib/pub_sub_model_sync/message_processor.rb
@@ -1,40 +1,54 @@
 # frozen_string_literal: true
 
 module PubSubModelSync
-  class MessageProcessor
-    attr_accessor :data, :klass, :action
+  class MessageProcessor < PubSubModelSync::Base
+    attr_accessor :payload
 
-    # @param data (Hash): any hash value to deliver
-    def initialize(data, klass, action)
-      @data = data
-      @klass = klass
-      @action = action
+    # @param payload (Payload): payload to be delivered
+    # @Deprecated: def initialize(data, klass, action)
+    def initialize(payload, klass = nil, action = nil)
+      @payload = payload
+      return if @payload.is_a?(Payload)
+
+      # support for deprecated
+      log('Deprecated: Use Payload instead of new(data, klass, action)')
+      @payload = PubSubModelSync::Payload.new(payload, { klass: klass, action: action })
     end
 
     def process
-      subscribers = filter_subscribers
-      subscribers.each { |subscriber| run_subscriber(subscriber) }
+      filter_subscribers.each(&method(:run_subscriber))
     end
 
     private
 
     def run_subscriber(subscriber)
-      subscriber.eval_message(data)
-      log "processed message with: #{[klass, action, data]}"
+      return unless processable?(subscriber)
+
+      subscriber.process!(payload)
+      res = config.on_success_processing.call(payload, subscriber)
+      log "processed message with: #{payload.inspect}" if res != :skip_log
     rescue => e
-      info = [klass, action, data, e.message, e.backtrace]
-      log("error processing message: #{info}", :error)
+      print_subscriber_error(e)
+    end
+
+    def processable?(subscriber)
+      cancel = config.on_before_processing.call(payload, subscriber) == :cancel
+      log("process message cancelled: #{payload}") if cancel && config.debug
+      !cancel
+    end
+
+    # @param error (Error)
+    def print_subscriber_error(error)
+      info = [payload, error.message, error.backtrace]
+      res = config.on_error_processing.call(error, payload)
+      log("Error processing message: #{info}", :error) if res != :skip_log
     end
 
     def filter_subscribers
-      PubSubModelSync::Config.subscribers.select do |subscriber|
-        subscriber.settings[:from_klass].to_s == klass.to_s &&
-          subscriber.settings[:from_action].to_s == action.to_s
+      config.subscribers.select do |subscriber|
+        subscriber.settings[:from_klass].to_s == payload.klass.to_s &&
+          subscriber.settings[:from_action].to_s == payload.action.to_s
       end
-    end
-
-    def log(message, kind = :info)
-      PubSubModelSync::Config.log message, kind
     end
   end
 end