+++ lib/pub_sub_model_sync/message_publisher.rb
@@ -1,17 +1,15 @@
 # frozen_string_literal: true
 
 module PubSubModelSync
-  class MessagePublisher
+  class MessagePublisher < PubSubModelSync::Base
     class << self
-      delegate :publish, to: :connector
-
       def connector
         @connector ||= PubSubModelSync::Connector.new
       end
 
       def publish_data(klass, data, action)
-        attrs = { klass: klass.to_s, action: action.to_sym }
-        publish(data, attrs)
+        payload = PubSubModelSync::Payload.new(data, { klass: klass.to_s, action: action.to_sym })
+        publish(payload)
       end
 
       # @param model: ActiveRecord model
@@ -22,11 +20,32 @@ module PubSubModelSync
 
         publisher ||= model.class.ps_publisher(action)
         payload = publisher.payload(model, action)
-        res_before = model.ps_before_sync(action, payload[:data])
+        res_before = model.ps_before_sync(action, payload.data)
         return if res_before == :cancel
 
-        publish(payload[:data], payload[:attrs])
-        model.ps_after_sync(action, payload[:data])
+        publish(payload)
+        model.ps_after_sync(action, payload.data)
+      end
+
+      def publish(payload)
+        if config.on_before_publish.call(payload) == :cancel
+          log("Publish message cancelled: #{payload}") if config.debug
+          return
+        end
+
+        log("Publishing message: #{[payload]}")
+        connector.publish(payload)
+        config.on_after_publish.call(payload)
+      rescue => e
+        notify_error(e, payload)
+      end
+
+      private
+
+      def notify_error(exception, payload)
+        info = [payload, exception.message, exception.backtrace]
+        res = config.on_error_publish.call(exception, payload)
+        log("Error publishing: #{info}", :error) if res != :skip_log
       end
     end
   end