+++ lib/pub_sub_model_sync/service_google.rb
@@ -7,10 +7,9 @@ end
 
 module PubSubModelSync
   class ServiceGoogle < ServiceBase
-    attr_accessor :service, :topic, :subscription, :config, :subscriber
+    attr_accessor :service, :topic, :subscription, :subscriber
 
     def initialize
-      @config = PubSubModelSync::Config
       @service = Google::Cloud::Pubsub.new(project: config.project,
                                            credentials: config.credentials)
       @topic = service.topic(config.topic_name) ||
@@ -28,13 +27,8 @@ module PubSubModelSync
       log('Listener stopped')
     end
 
-    def publish(data, attributes)
-      log("Publishing message: #{[attributes, data]}")
-      payload = { data: data, attributes: attributes }.to_json
-      topic.publish(payload, { SERVICE_KEY => true })
-    rescue => e
-      info = [attributes, data, e.message, e.backtrace]
-      log("Error publishing: #{info}", :error)
+    def publish(payload)
+      topic.publish(payload.to_json, { SERVICE_KEY => true })
     end
 
     def stop
@@ -45,23 +39,15 @@ module PubSubModelSync
     private
 
     def subscribe_to_topic
-      topic.subscription(config.subscription_name) ||
-        topic.subscribe(config.subscription_name)
+      topic.subscription(config.subscription_key) ||
+        topic.subscribe(config.subscription_key)
     end
 
     def process_message(received_message)
       message = received_message.message
-      return unless message.attributes[SERVICE_KEY]
-
-      perform_message(message.data)
-    rescue => e
-      log("Error processing message: #{[received_message, e.message]}", :error)
+      super(message.data) if message.attributes[SERVICE_KEY]
     ensure
       received_message.acknowledge!
-    end
-
-    def log(msg, kind = :info)
-      config.log("Google Service ==> #{msg}", kind)
     end
   end
 end