+++ spec/service_google_spec.rb
@@ -2,9 +2,9 @@
 
 RSpec.describe PubSubModelSync::ServiceGoogle do
   let(:msg_attrs) { { 'service_model_sync' => true } }
-  let(:message_data) { { data: {}, attributes: {} } }
+  let(:payload) { PubSubModelSync::Payload.new({}, {}) }
   let(:mock_message) do
-    double('Message', data: message_data.to_json, attributes: msg_attrs)
+    double('Message', data: payload.to_json, attributes: msg_attrs)
   end
   let(:mock_service_message) do
     double('ServiceMessage', message: mock_message, acknowledge!: true)
@@ -17,29 +17,29 @@ RSpec.describe PubSubModelSync::ServiceGoogle do
   before { allow(inst).to receive(:sleep) }
 
   describe 'initializer' do
-    it 'connect to pub/sub service' do
+    it 'connects to pub/sub service' do
       expect(inst.service).not_to be_nil
     end
-    it 'connect to topic' do
+    it 'connects to topic' do
       expect(inst.topic).not_to be_nil
     end
   end
 
   describe '.listen_messages' do
-    it 'subscribe to topic' do
+    it 'subscribes to topic' do
       expect(inst.topic).to receive(:subscription)
       inst.listen_messages
     end
-    it 'subscription listen for new messages' do
+    it 'listens for new messages' do
       expect(inst.topic.subscription).to receive(:listen).and_call_original
       inst.listen_messages
     end
-    it 'start subscriber' do
+    it 'starts subscriber' do
       subscriber = inst.topic.subscription.listen
       expect(subscriber).to receive(:start)
       inst.listen_messages
     end
-    it 'wait for messages' do
+    it 'awaits for messages' do
       expect(inst).to receive(:sleep)
       inst.listen_messages
     end
@@ -47,15 +47,16 @@ RSpec.describe PubSubModelSync::ServiceGoogle do
 
   describe '.process_message' do
     let(:message_processor) { PubSubModelSync::MessageProcessor }
-    it 'ignore if not a pub/sub model sync message (unknown)' do
+    before { allow(inst).to receive(:log) }
+    it 'ignores if not a pub/sub model sync message (unknown)' do
       expect(message_processor).not_to receive(:new)
       inst.send(:process_message, mock_service_unknown_message)
     end
-    it 'process message' do
+    it 'sends payload to message processor' do
       expect(message_processor).to receive(:new).and_call_original
       inst.send(:process_message, mock_service_message)
     end
-    it 'error processing' do
+    it 'prints error processing when failed' do
       error_msg = 'Invalid params'
       allow(message_processor).to receive(:new).and_raise(error_msg)
       expect(inst).to receive(:log).with(include(error_msg), :error)
@@ -63,17 +64,17 @@ RSpec.describe PubSubModelSync::ServiceGoogle do
     end
 
     describe 'mark as received message' do
-      it 'when success' do
+      it 'marks message as received when success' do
         expect(mock_service_message).to receive(:acknowledge!)
         inst.send(:process_message, mock_service_message)
       end
-      it 'when error' do
+      it 'marks message as received even if failed' do
         expect(mock_service_message).to receive(:acknowledge!)
         allow(mock_message).to receive(:data).and_return('invalid_data')
         allow(inst).to receive(:log)
         inst.send(:process_message, mock_service_message)
       end
-      it 'when unknown message' do
+      it 'marks message as received even if unknown message' do
         expect(mock_service_unknown_message).to receive(:acknowledge!)
         inst.send(:process_message, mock_service_unknown_message)
       end
@@ -82,11 +83,8 @@ RSpec.describe PubSubModelSync::ServiceGoogle do
 
   describe '.publish' do
     it 'delivery message' do
-      data = { name: 'test' }
-      attrs = { id: 10 }
-      payload = { data: data, attributes: attrs }.to_json
-      expect(inst.topic).to receive(:publish).with(payload, anything)
-      inst.publish(data, attrs)
+      expect(inst.topic).to receive(:publish).with(payload.to_json, anything)
+      inst.publish(payload)
     end
   end